+++ src/ts/modules/kanjiHover.ts
@@ -1,8 +1,8 @@
 import { RunnableAsyncModule } from '../module';
 import { getOption } from '../options';
-import { selectPersist } from '../spersist';
-import { plainToRuby, CARD_KEY, filterCards } from '../utils';
-import { getFieldValue, Field, cacheFieldValue } from '../fields';
+import { SPersistInterface, selectPersist } from '../spersist';
+import { plainToRuby, getCardKey, filterCards } from '../utils';
+import { getFieldValue, getFieldValueEle, Field, cacheFieldValue } from '../fields';
 
 import {
   Tooltips,
@@ -22,6 +22,9 @@ import {
   cardIDsToCardsInfo,
 } from '../ankiConnectUtils';
 import { translatorStrs } from '../consts';
+import {getViewportWidth} from '../reflow';
+import {compileOpts} from '../consts';
+import {MobilePopup} from '../mobilePopup';
 
 type KanjiHoverCategoryID =
   | 'word.nonNew.hidden'
@@ -44,14 +47,14 @@ type KanjiToFilteredCardIDs = Record<string, FilteredCardIDs>;
 type FilteredCardsInfo = Record<FilteredCardCategories, CardInfo[]>;
 type KanjiToFilteredCardsInfo = Record<string, FilteredCardsInfo>;
 
-type NoteInfoKanjiHover = {
+type KanjiToHoverHTML = Record<string, string>;
+
+export type NoteInfoKanjiHover = {
   WordReadingRubyHTML: string;
   WordReading: string;
   Key: string;
 };
 
-const kanjiHoverCardCacheKey = 'KanjiHover.kanjiHoverCardCacheKey';
-const kanjiHoverCardMutexKey = "KanjiHover.kanjiHoverCardMutexKey";
 
 // maps a kanji to the full hover HTML (the html containing the kanji + the full popup)
 type KanjiToHover = Record<string, string>;
@@ -60,18 +63,31 @@ export class KanjiHover extends RunnableAsyncModule {
   private readonly persist = selectPersist();
   private readonly persistObj = selectPersist('window');
   private readonly tooltips: Tooltips;
-  private readonly cardCacheKey: string;
-  private readonly mutexKey: string;
+  private readonly cardWordReadingResultKey: string;
+  private readonly cardHoverHTMLCacheKey: string;
+  private readonly mobilePopup: MobilePopup | null;
+  private readonly wordReadingEle: HTMLElement | null;
 
-  constructor() {
+  private readonly noteInfo: NoteInfoKanjiHover = {
+    WordReading: getFieldValue('WordReading'),
+    WordReadingRubyHTML: plainToRuby(getFieldValue('WordReading')),
+    Key: getFieldValue('Key'),
+  } as const;
+
+  constructor(mobilePopup: MobilePopup | null) {
     super('kanjiHover');
-
-    this.cardCacheKey = `${kanjiHoverCardCacheKey}.${CARD_KEY}`
-    this.mutexKey = `${kanjiHoverCardMutexKey}.${CARD_KEY}`;
 
     this.tooltips = new Tooltips();
     this.tooltips.overrideOptions(getOption('kanjiHover.overrideOptions.tooltips'));
 
+    this.cardWordReadingResultKey = `KanjiHover.cardWordReadingResultKey.${getCardKey()}`
+    //this.wordKanjisCacheKey = `KanjiHover.wordKanjisCacheKey.${getCardKey()}`;
+    this.cardHoverHTMLCacheKey = `KanjiHover.cardHoverHTMLCacheKey.${getCardKey()}`;
+    //this.mutexKey = `KanjiHover.mutexKey.${getCardKey()}`;
+
+    this.mobilePopup = mobilePopup;
+
+    this.wordReadingEle = document.getElementById('dh_reading');
 
     // ran synchronously, so fields will 100% be cached
     const cacheFields: readonly Field[] = [
@@ -84,6 +100,10 @@ export class KanjiHover extends RunnableAsyncModule {
     }
 
   }
+
+  //private getCardKanjiCacheKey(kanji: string) {
+  //  return this.cardWordReadingResultKey + "." + kanji;
+  //}
 
   private wordReadingKanjis(noteInfo: NoteInfoKanjiHover): Set<string> {
     let kanjiSet = new Set<string>(); // set of kanjis that requires api calls
@@ -408,16 +428,15 @@ export class KanjiHover extends RunnableAsyncModule {
     return tooltipBuilder.build().outerHTML;
   }
 
-  private async getKanjisToHover(
+  async getKanjisToHover(
     noteInfo: NoteInfoKanjiHover
   ): Promise<Record<string, string>> {
+    // Nothing is cached in this function directly, caching happens in
+    // the ankiConnectUtils queries / cardsInfo
     let kanjiSet = this.wordReadingKanjis(noteInfo);
 
     // maps a kanji to the full hover HTML (the html containing the kanji + the full popup)
     let kanjiToHover: Record<string, string> = {};
-
-    // looks for a cached hoverHTMLKey that doesn't contain the target word
-    // Array.from is to shallow-copy, so it doesn't interfere with kanjiSet.delete()
 
     // searches the remaining kanjis in kanjiSet
     const queryResults = await this.cardQueries(noteInfo, Array.from(kanjiSet));
@@ -443,79 +462,142 @@ export class KanjiHover extends RunnableAsyncModule {
     return kanjiToHover;
   }
 
-  async populateTooltips() {
-    if (this.persist?.has(this.mutexKey)) {
-      this.logger.debug("Cannot run due to mutex");
-      return;
-    }
-    this.logger.debug(`Running on ${CARD_KEY}...`);
-    this.persist?.set(this.mutexKey, "true")
+  private getWordReadingResult(kanjiToHoverHTML: KanjiToHoverHTML, wordReadingRubyHTML: string): string {
 
-    let wordReadingEle = document.getElementById('dh_reading');
-
-    const noteInfo: NoteInfoKanjiHover = {
-      WordReading: getFieldValue('WordReading'),
-      WordReadingRubyHTML: plainToRuby(getFieldValue('WordReading')),
-      Key: getFieldValue('Key'),
-    };
-
-    // checks for cache per kanji character
-    let kanjiToHoverHTML = await this.getKanjisToHover(noteInfo);
-
-
-    let resultHTML: string = noteInfo.WordReadingRubyHTML;
+    let resultHTML = wordReadingRubyHTML;
     if (Object.keys(kanjiToHoverHTML).length > 0) {
       // an empty object seems to replace every empty space
       const re = new RegExp(Object.keys(kanjiToHoverHTML).join('|'), 'gi');
-      resultHTML = noteInfo.WordReadingRubyHTML.replace(re, function (matched) {
+      resultHTML = wordReadingRubyHTML.replace(re, function (matched) {
         return `<span data-kanji-hover="${matched}">${kanjiToHoverHTML[matched]}</span>`;
       });
+    }
+    return resultHTML;
+  }
 
-      // can be null because this can be ran at the front side of the card
-      if (wordReadingEle !== null) {
-        wordReadingEle.innerHTML = resultHTML;
-        this.tooltips.addBrowseOnClick(wordReadingEle);
-      }
+  // attempts to get things without cache
+  async populateTooltips() {
+    //if (this.persist?.has(this.mutexKey)) {
+    //  this.logger.debug("Cannot run due to mutex");
+    //  return;
+    //}
+    if (this.wordReadingEle === null) {
+      this.logger.debug("Cannot run kanji hover due to null wordReadingEle");
+      return;
+    }
+    this.logger.debug(`Running on ${getCardKey()}...`);
+    //this.persist?.set(this.mutexKey, "true")
+
+    // gets popup for each kanji character and cache
+    let kanjiToHoverHTML = await this.getKanjisToHover(this.noteInfo);
+    this.displayResultFromKanjiToHoverHTML(kanjiToHoverHTML, this.wordReadingEle)
+
+    //this.persist?.pop(this.mutexKey)
+    this.logger.debug(`Finished running on ${getCardKey()}...`);
+  }
+
+  private displayResultFromKanjiToHoverHTML(kanjiToHoverHTML: KanjiToHoverHTML, wordReadingEle: HTMLElement) {
+
+    this.persist?.set(this.cardHoverHTMLCacheKey, JSON.stringify(kanjiToHoverHTML));
+
+    // this technically can be optimized depending on mobile or not, but imo
+    // it's safer to save this on both pc/mobile display
+    let resultHTML = this.getWordReadingResult(kanjiToHoverHTML, this.noteInfo.WordReadingRubyHTML)
+    this.persist?.set(this.cardWordReadingResultKey, resultHTML); // caches final result per card
+
+    // mobilePopup is not null <=> bp < combinePicture <=> should display mobile
+    if (this.mobilePopup !== null) {
+      this.mobilePopup.displayKanjiHover(kanjiToHoverHTML, wordReadingEle, this.noteInfo.WordReadingRubyHTML);
+    } else { // display normally
+      wordReadingEle.innerHTML = resultHTML;
+      this.tooltips.addBrowseOnClick(wordReadingEle);
+    }
+  }
+
+  private displayCachedResult(persist: SPersistInterface, wordReadingEle: HTMLElement) {
+    this.logger.debug("Using cached card");
+
+    // mobilePopup is not null <=> bp < combinePicture <=> should display mobile
+    if (this.mobilePopup !== null) {
+      const kanjiToHoverHTML: KanjiToHoverHTML = JSON.parse(persist.get(this.cardHoverHTMLCacheKey));
+
+      // construct kanjiToHoverHTML from cache
+      //const kanjiToHoverHTML: Record<string, string> = {}
+
+      //const kanjis = persist.get(this.wordKanjisCacheKey);
+      //if (kanjis === null) {
+      //  this.logger.warn("Cannot find kanjis?")
+      //  return;
+      //}
+
+      //for (const kanji of kanjis) {
+      //  const hoverResult = persist.get(this.getCardKanjiCacheKey(kanji));
+      //  if (hoverResult === null) {
+      //    this.logger.warn(`Skipping un-cached kanji ${kanji}...`)
+      //  } else {
+      //    kanjiToHoverHTML[kanji] = hoverResult;
+      //  }
+      //}
+
+      this.mobilePopup.displayKanjiHover(kanjiToHoverHTML, wordReadingEle, this.noteInfo.WordReadingRubyHTML);
+      // TODO addBrowseOnClick on mobile?
+
+    } else { // display normally
+      const resultHTML = persist.get(this.cardWordReadingResultKey);
+      wordReadingEle.innerHTML = resultHTML;
+      this.tooltips.addBrowseOnClick(wordReadingEle);
     }
 
-    // caches card
-    if (this.persist !== null) {
-      this.persist.set(this.cardCacheKey, resultHTML);
-    }
-
-    this.persist?.pop(this.mutexKey)
-    this.logger.debug(`Finished running on ${CARD_KEY}...`);
   }
 
   async main() {
+
+    if (this.useCache && this.wordReadingEle !== null) {
+      // checks for CardCache field first
+      const cardCacheEle = getFieldValueEle("CardCache");
+      const kanjiHoverData = cardCacheEle?.querySelector(`[data-cache-type="kanji-hover"]`);
+      if (kanjiHoverData) {
+        // format said data into kanjiToHoverHTML
+        const kanjiToHoverHTML: KanjiToHoverHTML = {};
+        for (const child of kanjiHoverData.children) {
+          const kanji = child.getAttribute("data-cache-kanji");
+          const hoverHTML = child.innerHTML;
+          if (kanji !== null) {
+            kanjiToHoverHTML[kanji] = hoverHTML;
+          }
+        }
+
+        this.logger.debug("Using CardCache");
+        console.log(kanjiToHoverHTML);
+        this.displayResultFromKanjiToHoverHTML(kanjiToHoverHTML, this.wordReadingEle);
+        return;
+      }
+
+      // checks for standard persist after
+      if (this.persist !== null && this.persist.has(this.cardWordReadingResultKey)) {
+        this.displayCachedResult(this.persist, this.wordReadingEle);
+        return;
+      }
+    }
 
     if (!this.getOption("enableAnkiconnectFeatures")) {
       return;
     }
 
-    let wordReadingEle = document.getElementById('dh_reading');
-
-    if (this.useCache && wordReadingEle !== null && this.persist !== null && this.persist.has(this.cardCacheKey)) {
-      this.logger.debug("Using cached card");
-      const resultHTML = this.persist.get(this.cardCacheKey);
-      wordReadingEle.innerHTML = resultHTML;
-      this.tooltips.addBrowseOnClick(wordReadingEle);
-      return;
-    }
-
     if (this.getOption("kanjiHover.activateOn") === "hover") {
-      if (wordReadingEle !== null) {
+      if (this.wordReadingEle !== null) {
 
-        wordReadingEle.onmouseover = (() => {
+        this.wordReadingEle.onmouseover = (() => {
           // replaces the function with a null function to avoid calling this function
-          if (wordReadingEle !== null) {
-            wordReadingEle.onmouseover = function() {}
+          if (this.wordReadingEle !== null) {
+            this.wordReadingEle.onmouseover = function() {}
           }
           this.populateTooltips();
         });
       } // otherwise front side. we do nothing! (pre-loading will lag the card)
     } else {
-      await this.populateTooltips();
+      // TODO: remove the above option! It should't be an option in the first place!
+      //await this.populateTooltips();
     }
   }
 }