+++ tools/make.py
@@ -17,11 +17,12 @@ from jinja2 import (
     select_autoescape,
     StrictUndefined,
 )
-from json_minify import json_minify
-import json
+#from json_minify import json_minify
+#import json
 
 import utils
 from note_changes import NOTE_CHANGES
+from json_handler import JsonHandler
 
 
 FRONT_FILENAME = "front.html"
@@ -173,6 +174,7 @@ class Generator:
         self,
         jinja_root_folders: list[str],
         config: utils.Config,
+        json_handler: JsonHandler,
         to_release: bool = False,
     ):
         self.jinja_root_folders = jinja_root_folders
@@ -183,6 +185,7 @@ class Generator:
             undefined=StrictUndefined,
             extensions=["jinja2.ext.do"],
         )
+        self.json_handler = json_handler
 
         filters = {
             # https://eengstrom.github.io/musings/add-bitwise-operations-to-ansible-jinja2
@@ -208,8 +211,9 @@ class Generator:
             root_folder, config("translation-file").item()
         )
 
-        with open(translation_file_path, encoding="utf-8") as f:
-            translations = json.loads(json_minify(f.read()))
+        translations = self.json_handler.read_file(translation_file_path)
+        #with open(translation_file_path, encoding="utf-8") as f:
+        #    translations = self.json_handler.read_file(f)
 
         languages = compile_options("displayLanguages").list()
         translator = Translator(languages, translations)
@@ -219,7 +223,7 @@ class Generator:
 
         self.data = {
             # helper methods
-            "NOTE_FILES": utils.get_note_config(),
+            "NOTE_DATA": utils.get_note_data(self.json_handler),
             # TODO change this to be based off of whatever version you specify
             "ALL_FIELDS": NOTE_CHANGES[0].fields,
             "COMPILE_OPTIONS": utils.get_compile_opts(config),
@@ -356,8 +360,7 @@ class Generator:
             utils.gen_dirs(release_output)
             shutil.copy(output_file, release_output)
 
-
-def create_generator(args: argparse.Namespace, config: utils.Config):
+def create_generator(args: argparse.Namespace, config: utils.Config, json_handler: JsonHandler):
     # search folders are: override, theme, src
     root_folder = utils.get_root_folder()
     templates_folder = os.path.join(root_folder, "src")
@@ -374,6 +377,7 @@ def create_generator(args: argparse.Namespace, config: utils.Config):
     generator = Generator(
         search_folders,
         config,
+        json_handler,
         to_release=args.to_release,
     )
     generator.set_data("VERSION", utils.get_version(args))
@@ -399,7 +403,7 @@ def generate_ts_consts(args: argparse.Namespace, generator: Generator):
 
 def generate_cards(args: argparse.Namespace, generator: Generator):
     root_folder = utils.get_root_folder()
-    note_config = utils.get_note_config()
+    note_data = utils.get_note_data(generator.json_handler)
 
 
     # generates typescript
@@ -426,8 +430,8 @@ def generate_cards(args: argparse.Namespace, generator: Generator):
         exit(e)
 
     # generates for each card type
-    note_model_id = note_config("id").item()
-    for card_model_id in note_config("templates").dict():
+    note_model_id = note_data("id").item()
+    for card_model_id in note_data("templates").dict():
         for side in ("front", "back"):
             file_name = side + ".html"
             input_file = os.path.join(note_model_id, card_model_id, file_name)
@@ -440,26 +444,12 @@ def generate_cards(args: argparse.Namespace, generator: Generator):
                 utils.Config({
                     "card-side": side,
                     "card-type": card_model_id,
-                    "card-type-name": note_config("templates", card_model_id, "name").item(),
-                    "model-name": note_config("model-name").item(),
+                    "card-type-name": note_data("templates", card_model_id, "name").item(),
+                    "model-name": note_data("model-name").item(),
                     "note-type": note_model_id,
-                    "js-prefix": note_config("js-prefix").item(),
+                    "js-prefix": note_data("js-prefix").item(),
                 }),
             )
-
-            # generates typescript for each model
-            #build_file(
-            #    args,
-            #    generator,
-            #    utils.Config(
-            #        {
-            #            "input-file": "ts/consts.ts.template",
-            #            "output-file": "tmp/ts/consts.ts",
-            #            "type": "jinja",
-            #            "to-release": False,
-            #        },
-            #    ),
-            #)
 
             generator.generate(
                 GenerateType.JINJA,
@@ -507,8 +497,9 @@ def main(args=None):
     if args.release:
         args.to_release = True
 
-    config = utils.get_config(args)
-    generator = create_generator(args, config)
+    json_handler = utils.create_json_handler(args)
+    config = utils.get_config(args, json_handler)
+    generator = create_generator(args, config, json_handler)
 
     if args.dev_generate_consts:
         generate_ts_consts(args, generator)
@@ -517,8 +508,8 @@ def main(args=None):
     generate_cards(args, generator)
 
     # generates each file in "build"
-    note_config = utils.get_note_config()
-    for file_config in note_config("build").list_items():
+    note_data = utils.get_note_data(json_handler)
+    for file_config in note_data("build").list_items():
         build_file(args, generator, file_config)